calculate Capacity Plan and new Capacity Requirements

+++ src/main/java/com/ibsys/backend/core/repository/WaitinglistRepository.java
@@ -4,6 +4,9 @@ import com.ibsys.backend.core.domain.entity.Waitinglist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface WaitinglistRepository extends JpaRepository<Waitinglist, Long> {
+public interface WaitinglistRepository extends JpaRepository<Waitinglist, Integer> {
+
 }
